Add interfaces for certificate registration and listing
Initial version of certificate registration form
allowing authorized users to issue new certificates
to a licensee by either a listing of number or provided
range.

Prototype version of certificate listing/search interface.
Users can view and search all certificates which they are
authorized to access.

Bringing in datatables.js and associated django-datatables-view
package to leverage the out-of-the-box support for paging, sorting,
and server-side processing.