Initial version of Licensee Address book
Licensee's are able to store commonly used destinations
(name, address, and country) for use when preparing certificates.

When preparing a certificate, a licensee user is able to select
from these stored destinations to pre-populate either the Exporter
or the Consignee fields. Both fields are editable after pre-populating.

The list of currently stored destinations is accessible and modifiable
via the Licensee page.

Only administrative users and a licensee's contacts are able to modify
an address book.