Faciliate and document website change workflow
Administrators of the USKPA website will
periodically need to modify the IRS Form 990
documents displayed on the website. To minimize
the effort required for this change, we inventory
and display links to, all files placed within the
uskpa_documents/irs folder. Allowing these changes
to be effected with a single commit.

Include documentation detailing the basic
git workflow and deployment process
for making a change such as this to a
deployed instance of the USKPA website.